Dynamic Tab Navigation Support (#6)
* CFCW-6 Initial commit

* Added App Navigation Content Item type

Added a type of Content Item that is considered "App Navigation" if the Item contains another Content Channel inside of it. Right now, the Content Channel that's inside the Item needs to be explicitly defined as "Item Content Channel"

* Mobile App Navigation Channel and Configuration

Continued development on creating a Content Channel that has an attribute that is a Content Channel. This is interpretted as an App Navigation Content Item. Yml Config also has a spot for setting the Content Channel that is to be used by the App to pull in Navigation

* API Support

Extended Query for getting the app navigation items based on an ID for a Rock Content Channel (configured in config.yml)

* Query Support for Dynamic Navigation

Added support for the navigation to be dynamically added based on a query return value.

Co-Authored-By: Vincent Wilson <vinnyjth@users.noreply.github.com>

* Added color support

Colors for the active tab color can be set dynamically

* CFCW-26 | Font Awesome 5

Updated the project to use Font Awesome 5 as a font library for both Android and iOS

* Updated Tab Object

Updated the Tab Object to use the createStackNavigator so navigationOptions get passed to allow for the header to persist

* Properly link icons and add fix navigating to connect tab

* Move header options into a single file for now

* Cleanign up unecessary code

* Fixed android

* Small linter cleanup

* Removed .env from git index

* Add .env.production to gitignore for now