App configuration updates (#11)
* Android Bundler ID Update

Updated the Android Bundle ID to io.echurch.christfellowshipconference so that we can replace the current Android App

* iOS Updates

Updated Version Number and there was a package that wasn't added that needed to be. This commit may not be necessary, but better safe than sorry

* Android Visual Assets

Updated the Splash Screens and App Icons for Android. The icons that I created are not curved in the edges, so we may have to revisit this later on, but I just wanted proper branding for the time being.

* Profile screen updates (#12)

* CFCW-32 Removed Liked Metrics

* CFCW-56 | Added new Auth type to handle first name and last name uploads on registration

* Expanded the getCurrentPerson Return Attributes to allow for custom data attributes

* Updated custom Person

Updated with better file organization and extended Rock Data Connector Person dataSource instead of redefining it.

* CFCW-14 | Visual Updates to Profile Screen

* Make sure we are including non-People resolvers (specifically mutation) when extending People

* Fix connect screen header

* Updated updateCurrentUser query

Updated the query to allow for attributeValues to be passed in as used for the Church, Department, and Job Title

* Feed fixes + Cloudinary (#13)

* Introduce subfeeds

* Added cloudinary

* Whats coming next (#14)

* Updated Content Items Query

Updated the Content Items file organization and fixed Time Zone error on the byUserFeed method

* iOS Version Update

* Added Coming Next View

* Updated Header Profile Icon

Updated the Profile Icon to be darker to compensate for the Cyan Header

* Added Moment.js

Added Moment.js and updated the LIVE_CONTENT method to use moment as the Timezone resolver

* Fix tapping into people on people feed